Choice of Means of Transport

This is an artificial data set on choice of means of transport based on cost and walking distance.

Usage

data(Transport)

Format

A data frame containing the following variables:

transport

means of transportation that can be chosen.

suburb

identifying number for each suburb

distance

walking distance to bus or train station

cost

cost of each means of transportation

working

size of working population of each suburb

prop.true

true choice probabilities

resp

choice frequencies of means of transportation