Change baseline category of multinomial logit or similar model

`rebase` returns an model object that is equivalent to the one given as argument but differs in parameterization

Usage

rebase(object, to, ...)

# S3 method for class 'mblogit'
rebase(object, to, ...)

Arguments

object

a statistical model object

to

usually, a string; the baseline category

...

other arguments, currently ignored